@@ -1,32 +1,41 @@
 [package]
 name = "gamesman-nova"
 version = "0.1.5"
-description = "System facilitating the search of extensive-form games."
 authors = ["Max Fierro <maxfierro@berkeley.edu>"]
 license = "GPL-3.0"
 repository = "https://github.com/GamesCrafters/GamesmanNova"
 keywords = ["game", "solver", "strong-solver", "research"]
-edition = "2021"
+edition = "2024"
 readme = "README.md"
-exclude = ["/.github"]
+exclude = ["/.github", "/dev", "/doc"]
+description = """Research system generating datasets of sequential game states \
+and associated features."""
 
 [[bin]]
-name = "nova"
 path = "src/main.rs"
-test = true
+name = "nova"
 bench = true
+test = true
 
 [profile.release]
-lto = "fat"
 codegen-units = 1
 debug = true
+lto = "fat"
 
 [dependencies]
+sqlx = { version = "0.7", features = ["sqlite", "runtime-tokio", "macros"] }
 clap = { version = "^4", features = ["derive"] }
+tokio = { version = "1", features = ["full"] }
 serde_json = "^1"
+once_cell = "^1"
 exitcode = "^1"
-nalgebra = "^0"
-colored = "^2"
+dotenv = "^0"
 anyhow = "^1"
 bitvec = "^1"
 regex = "^1"
+
+[dev-dependencies]
+strum_macros = "0.26"
+serial_test = "2.0"
+petgraph = "^0"
+strum = "0.26"
