@@ -2,10 +2,6 @@
 //!
 //! This module defines possible errors that could happen as a result of user
 //! input or an incomplete game implementation.
-//!
-//! #### Authorship
-//!
-//! - Max Fierro, 2/24/2024 (maxfierro@berkeley.edu)
 
 use std::{error::Error, fmt};
 
@@ -16,38 +12,23 @@ use std::{error::Error, fmt};
 /// of interface elements in `crate::game::mod`.
 #[derive(Debug)]
 pub enum GameError {
-    /// An error to indicate that a user attempted to solve a game variant
-    /// which is valid, but has no solver available to solve it.
-    SolverNotFound { input_game_name: &'static str },
+    /// An error to indicate that the variant passed to the game with the name
+    /// `game` was not in a format the game could parse. Includes a message
+    /// from the game implementation on exactly what went wrong. Note: `game`
+    /// should be a valid argument to the `--target` parameter in the CLI.
+    VariantMalformed { game: &'static str, hint: String },
 
-    /// An error to indicate that the variant passed to the game with
-    /// `game_name` was not in a format the game could parse. Includes a
-    /// message from the game implementation on exactly what went wrong. Note
-    /// that `game_name` should be a valid argument to the `--target`
-    /// parameter in the CLI.
-    VariantMalformed {
-        game_name: &'static str,
-        hint: String,
-    },
-
-    /// An error to indicate that the state string passed to the game with
-    /// `game_name` was not in a format the game could parse. Includes a
-    /// message from the game implementation on exactly what went wrong. Note
-    /// that `game_name` should be a valid argument to the `--target`
-    /// parameter in the CLI.
-    StateMalformed {
-        game_name: &'static str,
-        hint: String,
-    },
+    /// An error to indicate that the state string passed to the game with the
+    /// name `game` was not in a format the game could parse. Includes a message
+    /// from the game implementation on exactly what went wrong. Note: `game`
+    /// should be a valid argument to the `--target` parameter in the CLI.    
+    StateMalformed { game: &'static str, hint: String },
 
     /// An error to indicate that a sequence of states in string form would
     /// be impossible to reproduce in real play. Includes a message from the
-    /// game implementation on exactly what went wrong. Note that `game_name`
+    /// game implementation on exactly what went wrong. Note: `target_name`
     /// should be a valid argument to the `--target` parameter in the CLI.
-    InvalidHistory {
-        game_name: &'static str,
-        hint: String,
-    },
+    InvalidHistory { game: &'static str, hint: String },
 }
 
 impl Error for GameError {}
@@ -55,36 +36,27 @@ impl Error for GameError {}
 impl fmt::Display for GameError {
     fn fmt(&self, f: &mut fmt::Formatter) -> fmt::Result {
         match self {
-            Self::SolverNotFound { input_game_name } => {
+            Self::VariantMalformed { game, hint } => {
                 write!(
                     f,
-                    "The variant you specified for the game {} has no solvers \
-                    associated with it.",
-                    input_game_name
+                    "{hint}\n\nMore information on how the game expects you to \
+                    format variant encodings can be found with 'nova info \
+                    {game}'.",
                 )
             },
-            Self::VariantMalformed { game_name, hint } => {
+            Self::StateMalformed { game, hint } => {
                 write!(
                     f,
-                    "{}\n\n\tMore information on how the game expects you to \
-                    format it can be found with 'nova info {} --output extra'.",
-                    hint, game_name
+                    "{hint}\n\nMore information on how the game expects you to \
+                    format state encodings can be found with 'nova info \
+                    {game}'.",
                 )
             },
-            Self::StateMalformed { game_name, hint } => {
+            Self::InvalidHistory { game, hint } => {
                 write!(
                     f,
-                    "{}\n\n\tMore information on how the game expects you to \
-                    format it can be found with 'nova info {} --output extra'.",
-                    hint, game_name
-                )
-            },
-            Self::InvalidHistory { game_name, hint } => {
-                write!(
-                    f,
-                    "{}\n\n\tMore information on the game's rules can be found \
-                    with 'nova info {} --output extra'.",
-                    hint, game_name
+                    "{hint}\n\nMore information on the game's rules can be \
+                    found with 'nova info {game}'.",
                 )
             },
         }
