@@ -1,224 +1,162 @@
-//! # Game Utilities Module
+//! # Game Utility Module
 //!
-//! This module provides some common utilities used in the implementation of
-//! more than a single game.
-//!
-//! #### Authorship
-//!
-//! - Max Fierro, 11/2/2023 (maxfierro@berkeley.edu)
+//! This module defines utilities used game implementations.
 
+use anyhow::bail;
 use anyhow::{Context, Result};
 
-use crate::{
-    game::error::GameError,
-    game::{DTransition, Legible, STransition},
-    model::{PlayerCount, State, Turn},
-    solver::MAX_TRANSITIONS,
-};
+use std::fmt::Display;
 
-/* TURN ENCODING */
-
-/// Minimally encodes turn information into the 64-bit integer `state` by
-/// shifting the integer in `state` just enough bits to allow `turn` to be
-/// expressed, where `turn` is upper-bounded by `player_count`.
-///
-/// For example, if `player_count` is 3, `state` is `0b00...01`, and we want to
-/// encode that it is player `2`'s turn (where players are 0-indexed), we would
-/// return `0b00...00111`, whereas if `player_count` was 2 we would return
-/// `0b00...0011`. This is because you need two bits to enumerate `{0, 1, 2}`,
-/// but only one to enumerate `{0, 1}`.
-pub fn pack_turn(state: State, turn: Turn, player_count: PlayerCount) -> State {
-    if player_count == 0 {
-        return state;
-    } else {
-        let turn_bits = Turn::BITS - (player_count - 1).leading_zeros();
-        (state << turn_bits) | (turn as State)
-    }
-}
-
-/// Given a state and a player count, determines the player whose turn it is by
-/// taking note of the integer in the rightmost bits of `state`. The number of
-/// bits considered turn information are determined by `player_count`. This is
-/// the inverse function of `pack_turn`.
-pub fn unpack_turn(
-    encoding: State,
-    player_count: PlayerCount,
-) -> (State, Turn) {
-    if player_count == 0 {
-        return (encoding, 0);
-    } else {
-        let turn_bits = Turn::BITS - (player_count - 1).leading_zeros();
-        let turn_mask = (1 << turn_bits) - 1;
-        let state = (encoding & !turn_mask) >> turn_bits;
-        let turn = (encoding & turn_mask) as usize;
-        (state, turn)
-    }
-}
+use crate::game::GameData;
+use crate::game::Information;
+use crate::game::State;
+use crate::game::{Codec, Implicit, error::GameError};
+use crate::interface::GameAttribute;
 
 /* STATE HISTORY VERIFICATION */
 
-/// Returns the latest state in a sequential `history` of state string encodings
-/// by verifying that the first state in the history is the same as the `game`'s
-/// start and that each state can be reached from its predecessor through the
-/// `game`'s transition function. If these conditions are not met, it returns an
-/// error message signaling the pair of states that are not connected by the
-/// transition function, with a reminder of the current game variant.
-pub fn verify_history_dynamic<G>(
-    game: &G,
+/// Verifies that the elements of `history` are a valid sequence of states under
+/// the rules of `target`, failing if this is not true.
+pub fn verify_state_history<const B: usize, G>(
+    target: &G,
     history: Vec<String>,
-) -> Result<State>
+) -> Result<State<B>>
 where
-    G: Legible<State> + DTransition<State>,
+    G: Information + Implicit<B> + Codec<B>,
 {
-    if let Some(s) = history.first() {
-        let mut prev = game.decode(s.clone())?;
-        if prev == game.start() {
-            for i in 1..history.len() {
-                let next = game.decode(history[i].clone())?;
-                let transitions = game.prograde(prev);
+    let history = sanitize_input(history);
+    if let Some((l, s)) = history.first() {
+        let mut prev = target
+            .decode(s.clone())
+            .context(format!("Failed to parse line #{l}."))?;
+        if prev == target.source() {
+            for h in history.iter().skip(1) {
+                let (l, s) = h.clone();
+                let next = target
+                    .decode(s)
+                    .context(format!("Failed to parse line #{l}."))?;
+                if target.sink(prev) {
+                    bail!(
+                        terminal_history_error(target, prev, next)?.context(
+                            format!(
+                                "Invalid state transition found at line #{l}.",
+                            ),
+                        )
+                    )
+                }
+                let transitions = target.adjacent(prev);
                 if !transitions.contains(&next) {
-                    return transition_history_error(game, prev, next);
+                    bail!(
+                        transition_history_error(target, prev, next)?.context(
+                            format!(
+                                "Invalid state transition found at line #{l}."
+                            ),
+                        )
+                    )
                 }
                 prev = next;
             }
             Ok(prev)
         } else {
-            start_history_error(game, game.start())
+            bail!(GameError::InvalidHistory {
+                game: G::info().name,
+                hint: format!(
+                    "The state history must begin with the starting state for \
+                    the provided game variant, which is {}.",
+                    target.encode(target.source())?
+                ),
+            })
         }
     } else {
-        empty_history_error(game)
+        bail!(GameError::InvalidHistory {
+            game: G::info().name,
+            hint: "State history must contain at least one state.".into(),
+        })
     }
 }
 
-/// Returns the latest state in a sequential `history` of state string encodings
-/// by verifying that the first state in the history is the same as the `game`'s
-/// start and that each state can be reached from its predecessor through the
-/// `game`'s transition function. If these conditions are not met, it returns an
-/// error message signaling the pair of states that are not connected by the
-/// transition function, with a reminder of the current game variant.
-pub fn verify_history_static<G>(game: &G, history: Vec<String>) -> Result<State>
+/// Enumerates lines and trims whitespace from input.
+fn sanitize_input(mut input: Vec<String>) -> Vec<(usize, String)> {
+    input
+        .iter_mut()
+        .enumerate()
+        .map(|(i, s)| (i, s.trim().to_owned()))
+        .filter(|(_, s)| !s.is_empty())
+        .collect()
+}
+
+/* HISTORY VERIFICATION ERRORS */
+
+fn transition_history_error<const B: usize, G>(
+    target: &G,
+    prev: State<B>,
+    next: State<B>,
+) -> Result<anyhow::Error>
 where
-    G: Legible<State> + STransition<State, MAX_TRANSITIONS>,
+    G: Information + Codec<B>,
 {
-    if let Some(s) = history.first() {
-        let mut prev = game.decode(s.clone())?;
-        if prev == game.start() {
-            for i in 1..history.len() {
-                let next = game.decode(history[i].clone())?;
-                let transitions = game.prograde(prev);
-                if !transitions.contains(&Some(next)) {
-                    return transition_history_error(game, prev, next);
-                }
-                prev = next;
-            }
-            Ok(prev)
-        } else {
-            start_history_error(game, game.start())
-        }
-    } else {
-        empty_history_error(game)
-    }
-}
-
-fn empty_history_error<G: Legible<State>>(game: &G) -> Result<State> {
-    Err(GameError::InvalidHistory {
-        game_name: game.info().name,
-        hint: format!("State history must contain at least one state."),
-    })
-    .context("Invalid game history.")
-}
-
-fn start_history_error<G: Legible<State>>(
-    game: &G,
-    start: State,
-) -> Result<State> {
-    Err(GameError::InvalidHistory {
-        game_name: game.info().name,
+    bail!(GameError::InvalidHistory {
+        game: G::info().name,
         hint: format!(
-            "The state history must begin with the starting state for this \
-            variant ({}), which is {}.",
-            game.info().variant,
-            game.encode(start)
+            "Transitioning from the state '{}' to the sate '{}' is illegal in \
+            the provided target variant.",
+            target.encode(prev)?,
+            target.encode(next)?,
         ),
     })
-    .context("Invalid game history.")
 }
 
-fn transition_history_error<G: Legible<State>>(
-    game: &G,
-    prev: State,
-    next: State,
-) -> Result<State> {
-    Err(GameError::InvalidHistory {
-        game_name: game.info().name,
+fn terminal_history_error<const B: usize, G>(
+    target: &G,
+    prev: State<B>,
+    next: State<B>,
+) -> Result<anyhow::Error>
+where
+    G: Information + Codec<B>,
+{
+    bail!(GameError::InvalidHistory {
+        game: G::info().name,
         hint: format!(
-            "Transitioning from the state '{}' to the sate '{}' is \
-            illegal in the current game variant ({}).",
-            game.encode(prev),
-            game.encode(next),
-            game.info().variant
+            "Transitioning from the state '{}' to the sate '{}' is illegal in \
+            the provided target variant, because '{}' is a terminal state.",
+            target.encode(prev)?,
+            target.encode(next)?,
+            target.encode(prev)?,
         ),
     })
-    .context("Invalid game history.")
 }
 
-/* TESTS */
+/* GAME DATA UTILITIES */
 
-#[cfg(test)]
-mod test {
-
-    use super::*;
-
-    /* TURN ENCODING TESTS */
-
-    #[test]
-    fn pack_turn_correctness() {
-        // Require three turn bits (8 players = {0b000, 0b001, ..., 0b111})
-        let player_count: Turn = 8;
-        // 5 in decimal
-        let turn: Turn = 0b0000_0101;
-        // 31 in decimal
-        let state: State = 0b0001_1111;
-        // 0b00...00_1111_1101 in binary = 0b[state bits][player bits]
-        assert_eq!(0b1111_1101, pack_turn(state, turn, player_count));
+impl Display for GameAttribute {
+    fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
+        let content = match self {
+            GameAttribute::VariantProtocol => "variant-protocol",
+            GameAttribute::VariantPattern => "variant-pattern",
+            GameAttribute::VariantDefault => "variant-default",
+            GameAttribute::StateProtocol => "state-protocol",
+            GameAttribute::StateDefault => "state-default",
+            GameAttribute::StatePattern => "state-pattern",
+            GameAttribute::Authors => "authors",
+            GameAttribute::About => "about",
+            GameAttribute::Name => "name",
+        };
+        write!(f, "{content}")
     }
+}
 
-    #[test]
-    fn unpack_turn_correctness() {
-        // Require six turn bits (players = {0b0, 0b1, ..., 0b100101})
-        let player_count: Turn = 38;
-        // 346 in decimal
-        let encoding: State = 0b0001_0101_1010;
-        // 0b00...00_0001_0101_1010 -> 0b00...00_0101 and 0b0001_1010, which
-        // means that 346 should be decoded to a state of 5 and a turn of 26
-        assert_eq!((5, 26), unpack_turn(encoding, player_count));
-    }
-
-    #[test]
-    fn unpack_is_inverse_of_pack() {
-        // Require two turn bits (players = {0b00, 0b01, 0b10})
-        let player_count: Turn = 3;
-        // 0b00...01 in binary
-        let turn: Turn = 2;
-        // 0b00...0111 in binary
-        let state: State = 7;
-        // 0b00...011101 in binary
-        let packed: State = pack_turn(state, turn, player_count);
-        // Packing and unpacking should yield equivalent results
-        assert_eq!((state, turn), unpack_turn(packed, player_count));
-
-        // About 255 * 23^2 iterations
-        for p in Turn::MIN..=255 {
-            let turn_bits = Turn::BITS - p.leading_zeros();
-            let max_state: State = State::MAX / ((1 << turn_bits) as u64);
-            let state_step = ((max_state / 23) + 1) as usize;
-            let turn_step = ((p / 23) + 1) as usize;
-
-            for s in (State::MIN..max_state).step_by(state_step) {
-                for t in (Turn::MIN..p).step_by(turn_step) {
-                    assert_eq!((s, t), unpack_turn(pack_turn(s, t, p), p));
-                }
-            }
+impl GameData {
+    pub fn find(&self, attribute: GameAttribute) -> &str {
+        match attribute {
+            GameAttribute::VariantProtocol => self.variant_protocol,
+            GameAttribute::VariantPattern => self.variant_pattern,
+            GameAttribute::VariantDefault => self.variant_default,
+            GameAttribute::StateProtocol => self.state_protocol,
+            GameAttribute::StateDefault => self.state_default,
+            GameAttribute::StatePattern => self.state_pattern,
+            GameAttribute::Authors => self.authors,
+            GameAttribute::About => self.about,
+            GameAttribute::Name => self.name,
         }
     }
 }
