@@ -3,27 +3,21 @@
 //! This module helps parse the a string encoding of a zero-by game state into
 //! a more efficient binary representation, performing a series of checks which
 //! partially ensure compatibility with a game variant.
-//!
-//! #### Authorship
-//!
-//! - Max Fierro, 11/2/2023 (maxfierro@berkeley.edu)
 
 use regex::Regex;
 
+use crate::game::Player;
+use crate::game::State;
 use crate::game::error::GameError;
-use crate::game::util::pack_turn;
-use crate::game::util::unpack_turn;
-use crate::game::zero_by::Session;
+use crate::game::zero_by::Elements;
 use crate::game::zero_by::NAME;
-use crate::model::State;
-use crate::model::Turn;
+use crate::game::zero_by::Session;
 
 /* ZERO-BY STATE ENCODING */
 
-pub const STATE_DEFAULT: &'static str = "10-0";
-pub const STATE_PATTERN: &'static str = r"^\d+-\d+$";
-pub const STATE_PROTOCOL: &'static str =
-"The state string should be two dash-separated positive integers without any \
+pub const STATE_DEFAULT: &str = "10-0";
+pub const STATE_PATTERN: &str = r"^\d+-\d+$";
+pub const STATE_PROTOCOL: &str = "The state string should be two dash-separated positive integers without any \
 decimal points. The first integer will indicate the amount of elements left to \
 remove from the set, and the second indicates whose turn it is to remove an \
 element. The first integer must be less than or equal to the number of initial \
@@ -42,22 +36,21 @@ pub fn parse_state(
 ) -> Result<State, GameError> {
     check_state_pattern(&from)?;
     let params = parse_parameters(&from)?;
-    let (from, turn) = check_param_count(&params)?;
-    check_variant_coherence(from, turn, &session)?;
-    let state = pack_turn(from, turn, session.players);
-    Ok(state)
+    let (elements, turn) = check_param_count(&params)?;
+    check_variant_coherence(elements, turn, session)?;
+    Ok(session.encode_state(turn, elements))
 }
 
 /* STATE STRING VERIFICATION */
 
 fn check_state_pattern(from: &String) -> Result<(), GameError> {
     let re = Regex::new(STATE_PATTERN).unwrap();
-    if !re.is_match(&from) {
+    if !re.is_match(from) {
         Err(GameError::StateMalformed {
-            game_name: NAME,
+            game: NAME,
             hint: format!(
-                "String does not match the pattern '{}'.",
-                STATE_PATTERN
+                "Input string '{from}' does not match the pattern \
+                '{STATE_PATTERN}'.",
             ),
         })
     } else {
@@ -65,23 +58,23 @@ fn check_state_pattern(from: &String) -> Result<(), GameError> {
     }
 }
 
-fn parse_parameters(from: &String) -> Result<Vec<u64>, GameError> {
+fn parse_parameters(from: &str) -> Result<Vec<u64>, GameError> {
     from.split('-')
         .map(|int_string| {
             int_string
                 .parse::<u64>()
                 .map_err(|e| GameError::StateMalformed {
-                    game_name: NAME,
-                    hint: format!("{}", e.to_string()),
+                    game: NAME,
+                    hint: e.to_string(),
                 })
         })
         .collect()
 }
 
-fn check_param_count(params: &Vec<u64>) -> Result<(State, Turn), GameError> {
+fn check_param_count(params: &[u64]) -> Result<(Elements, Player), GameError> {
     if params.len() != 2 {
         Err(GameError::StateMalformed {
-            game_name: NAME,
+            game: NAME,
             hint: format!(
                 "String contains {} integers, but needs to have exactly 2.",
                 params.len()
@@ -93,27 +86,26 @@ fn check_param_count(params: &Vec<u64>) -> Result<(State, Turn), GameError> {
 }
 
 fn check_variant_coherence(
-    from: State,
-    turn: Turn,
+    from: Elements,
+    turn: Player,
     session: &Session,
 ) -> Result<(), GameError> {
-    let (session_from, _) = unpack_turn(session.start, session.players);
-    if from > session_from {
+    if from > session.start_elems {
         Err(GameError::StateMalformed {
-            game_name: NAME,
+            game: NAME,
             hint: format!(
-                "Specified more starting elements ({}) than variant allows \
+                "Specified more starting elements ({from}) than variant allows \
                 ({}).",
-                from, session.start,
+                session.start_elems,
             ),
         })
     } else if turn >= session.players {
         Err(GameError::StateMalformed {
-            game_name: NAME,
+            game: NAME,
             hint: format!(
-                "Specified a turn ({}) too high for this ({}-player) game \
+                "Specified a turn ({turn}) too high for this ({}-player) game \
                 variant.",
-                turn, session.players,
+                session.players,
             ),
         })
     } else {
@@ -127,7 +119,7 @@ fn check_variant_coherence(
 mod test {
 
     use super::*;
-    use crate::game::{util::verify_history_dynamic, Game};
+    use crate::game::*;
 
     /* STATE STRING PARSING */
 
@@ -144,11 +136,11 @@ mod test {
 
     #[test]
     fn no_state_equals_default_state() {
-        let with_none = Session::initialize(None).unwrap();
-        let with_default = Session::initialize(None).unwrap();
+        let with_none = Session::default();
+        let with_default = Session::default();
 
         assert_eq!(
-            with_none.start,
+            with_none.start_state,
             parse_state(&with_default, STATE_DEFAULT.to_string()).unwrap()
         );
     }
@@ -163,9 +155,9 @@ mod test {
         let s6 = "7-".to_owned();
         let s7 = "11-0".to_owned();
 
-        fn f() -> Session {
+        fn f() -> Session<'static> {
             // 2-player 10-to-zero by 1 or 2
-            Session::initialize(None).unwrap()
+            Session::default()
         }
 
         assert!(parse_state(&f(), s1).is_err());
@@ -187,8 +179,8 @@ mod test {
         let s6 = "10-1".to_owned(); // <-- Impossible but well formed
         let s7 = "1-0".to_owned();
 
-        fn f() -> Session {
-            Session::initialize(None).unwrap()
+        fn f() -> Session<'static> {
+            Session::default()
         }
 
         assert!(parse_state(&f(), s1).is_ok());
@@ -201,7 +193,7 @@ mod test {
     }
 
     #[test]
-    fn compatible_variants_and_states_pass_checks() {
+    fn compatible_variants_and_states_pass_checks() -> Result<()> {
         let v1 = "50-10-12-1-4";
         let v2 = "5-100-6-2-7";
         let v3 = "10-200-1-5";
@@ -210,21 +202,18 @@ mod test {
         let s2 = "150-9".to_owned();
         let s3 = "200-0".to_owned();
 
-        fn f(v: &str) -> Session {
-            Session::initialize(Some(v.to_owned())).unwrap()
-        }
+        assert!(parse_state(&variant(v1)?, s1.clone()).is_ok());
+        assert!(parse_state(&variant(v1)?, s2.clone()).is_err());
+        assert!(parse_state(&variant(v1)?, s3.clone()).is_err());
 
-        assert!(parse_state(&f(v1), s1.clone()).is_ok());
-        assert!(parse_state(&f(v1), s2.clone()).is_err());
-        assert!(parse_state(&f(v1), s3.clone()).is_err());
+        assert!(parse_state(&variant(v2)?, s1.clone()).is_ok());
+        assert!(parse_state(&variant(v2)?, s2.clone()).is_err());
+        assert!(parse_state(&variant(v2)?, s3.clone()).is_err());
 
-        assert!(parse_state(&f(v2), s1.clone()).is_ok());
-        assert!(parse_state(&f(v2), s2.clone()).is_err());
-        assert!(parse_state(&f(v2), s3.clone()).is_err());
-
-        assert!(parse_state(&f(v3), s1.clone()).is_ok());
-        assert!(parse_state(&f(v3), s2.clone()).is_ok());
-        assert!(parse_state(&f(v3), s3.clone()).is_ok());
+        assert!(parse_state(&variant(v3)?, s1.clone()).is_ok());
+        assert!(parse_state(&variant(v3)?, s2.clone()).is_ok());
+        assert!(parse_state(&variant(v3)?, s3.clone()).is_ok());
+        Ok(())
     }
 
     /* GAME HISTORY VERIFICATION */
@@ -235,43 +224,102 @@ mod test {
         let i2 = vec!["10-0", "8-0", "7-0", "5-1"]; // Turns don't switch
         let i3 = vec!["10-1", "8-0", "7-1", "5-0"]; // Starting turn wrong
         let i4 = vec!["1-10", "0-9", "1-7", "0-5"]; // Turn and state switched
-        let i5 = vec!["10-0", "", "9-1", "", "8-0"]; // Empty states
-        let i6 = vec!["ten-zero", "nine-one"]; // Malformed
-        let i7: Vec<&str> = vec![]; // No history
-        let i8 = vec![""]; // Empty string
+        let i5 = vec!["ten-zero", "nine-one"]; // Malformed
+        let i6: Vec<&str> = vec![]; // No history
+        let i7 = vec![""]; // Empty string
 
-        assert!(verify_history_dynamic(&session(None), owned(i1)).is_err());
-        assert!(verify_history_dynamic(&session(None), owned(i2)).is_err());
-        assert!(verify_history_dynamic(&session(None), owned(i3)).is_err());
-        assert!(verify_history_dynamic(&session(None), owned(i4)).is_err());
-        assert!(verify_history_dynamic(&session(None), owned(i5)).is_err());
-        assert!(verify_history_dynamic(&session(None), owned(i6)).is_err());
-        assert!(verify_history_dynamic(&session(None), owned(i7)).is_err());
-        assert!(verify_history_dynamic(&session(None), owned(i8)).is_err());
+        assert!(
+            Session::default()
+                .forward(owned(i1))
+                .is_err()
+        );
+
+        assert!(
+            Session::default()
+                .forward(owned(i2))
+                .is_err()
+        );
+
+        assert!(
+            Session::default()
+                .forward(owned(i3))
+                .is_err()
+        );
+
+        assert!(
+            Session::default()
+                .forward(owned(i4))
+                .is_err()
+        );
+
+        assert!(
+            Session::default()
+                .forward(owned(i5))
+                .is_err()
+        );
+
+        assert!(
+            Session::default()
+                .forward(owned(i6))
+                .is_err()
+        );
+
+        assert!(
+            Session::default()
+                .forward(owned(i7))
+                .is_err()
+        );
     }
 
     #[test]
     fn verify_correct_default_zero_by_history_passes() {
-        let c1 = vec!["10-0", "8-1", "6-0", "4-1", "2-0", "0-1"];
-        let c2 = vec!["10-0", "8-1", "6-0", "4-1", "2-0"];
-        let c3 = vec!["10-0", "9-1", "7-0", "6-1"];
-        let c4 = vec!["10-0", "8-1", "6-0"];
-        let c5 = vec!["10-0", "9-1"];
-        let c6 = vec!["10-0"];
+        let c1 = vec!["10-0", "8-1", " ", "6-0", "4-1", "2-0", "0-1"];
+        let c2 = vec!["", "10-0", "8-1", "6-0", "4-1", "2-0"];
+        let c3 = vec!["10-0", "9-1", "", "", "7-0", "6-1"];
+        let c4 = vec!["10-0", "8-1", "6-0", "  "];
+        let c5 = vec!["10-0", " ", "9-1"];
+        let c6 = vec!["", "10-0", " "];
 
-        assert!(verify_history_dynamic(&session(None), owned(c1)).is_ok());
-        assert!(verify_history_dynamic(&session(None), owned(c2)).is_ok());
-        assert!(verify_history_dynamic(&session(None), owned(c3)).is_ok());
-        assert!(verify_history_dynamic(&session(None), owned(c4)).is_ok());
-        assert!(verify_history_dynamic(&session(None), owned(c5)).is_ok());
-        assert!(verify_history_dynamic(&session(None), owned(c6)).is_ok());
+        assert!(
+            Session::default()
+                .forward(owned(c1))
+                .is_ok()
+        );
+
+        assert!(
+            Session::default()
+                .forward(owned(c2))
+                .is_ok()
+        );
+
+        assert!(
+            Session::default()
+                .forward(owned(c3))
+                .is_ok()
+        );
+
+        assert!(
+            Session::default()
+                .forward(owned(c4))
+                .is_ok()
+        );
+
+        assert!(
+            Session::default()
+                .forward(owned(c5))
+                .is_ok()
+        );
+
+        assert!(
+            Session::default()
+                .forward(owned(c6))
+                .is_ok()
+        );
     }
 
     #[test]
-    fn verify_zero_by_history_compatibility() {
-        fn v() -> Option<String> {
-            Some(format!("8-200-30-70-15-1"))
-        }
+    fn verify_zero_by_history_compatibility() -> Result<()> {
+        let v = "8-200-30-70-15-1";
 
         let c1 = vec![
             "200-0", "185-1", "115-2", "114-3", "113-4", "83-5", "82-6",
@@ -282,29 +330,59 @@ mod test {
             "110-7", "80-0", "79-1",
         ];
 
-        assert!(verify_history_dynamic(&session(v()), owned(c1)).is_ok());
-        assert!(verify_history_dynamic(&session(v()), owned(c2)).is_ok());
+        assert!(
+            &variant(v)?
+                .forward(owned(c1))
+                .is_ok()
+        );
+
+        assert!(
+            &variant(v)?
+                .forward(owned(c2))
+                .is_ok()
+        );
 
         let i1 = vec!["200-0", "184-1", "115-2", "114-3"]; // Illegal move
         let i2 = vec!["200-0", "185-1", "115-1", "114-2"]; // Turns don't switch
         let i3 = vec!["200-2", "185-3", "115-4", "114-5"]; // Bad initial turn
         let i4 = vec!["201-0", "186-1", "116-2", "115-3"]; // Bad initial state
 
-        assert!(verify_history_dynamic(&session(v()), owned(i1)).is_err());
-        assert!(verify_history_dynamic(&session(v()), owned(i2)).is_err());
-        assert!(verify_history_dynamic(&session(v()), owned(i3)).is_err());
-        assert!(verify_history_dynamic(&session(v()), owned(i4)).is_err());
+        assert!(
+            &variant(v)?
+                .forward(owned(i1))
+                .is_err()
+        );
+
+        assert!(
+            &variant(v)?
+                .forward(owned(i2))
+                .is_err()
+        );
+
+        assert!(
+            &variant(v)?
+                .forward(owned(i3))
+                .is_err()
+        );
+
+        assert!(
+            &variant(v)?
+                .forward(owned(i4))
+                .is_err()
+        );
+
+        Ok(())
     }
 
     /* UTILITIES */
 
-    fn session(v: Option<String>) -> Session {
-        Session::initialize(v).unwrap()
+    fn variant(v: &str) -> Result<Session> {
+        Session::variant(v.to_string())
     }
 
-    fn owned(v: Vec<&str>) -> Vec<String> {
+    fn owned(v: Vec<&'static str>) -> Vec<String> {
         v.iter()
-            .map(|&s| s.to_owned())
+            .map(|s| s.to_string())
             .collect()
     }
 }
